requirements.txt
Pillow
colorama
keystone-engine
numpy
pycrypto
pyelftools
